admin profile - working user update record information - IV

+++ app/Traits/FileUploadTraits.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use File;
 
 trait FileUploadTraits {
-    function uploadImage(Request $request,string $inputName,string $oldPath=null,string $path= '/uploads') : ?string
+    function uploadImage(Request $request, string $inputName, ?string $oldPath=null, string $path= '/uploads') : ?string
     {
         if ($request->hasFile($inputName)) {
             $image = $request->{$inputName};
@@ -16,7 +16,8 @@ trait FileUploadTraits {
             $image->move(public_path($path), $imageName);
 
             //delete previous image from storage
-            if ($oldPath && File::exists(public_path($oldPath))) {
+            $exculudedFolder = '/default';
+            if ($oldPath && File::exists(public_path($oldPath)) && strpos($oldPath,$exculudedFolder) !== 0) {
                 File::delete(public_path($oldPath));
             }
